+++ braindecode/models/util.py
@@ -233,6 +233,7 @@ models_mandatory_parameters = [
     ("LMDANet", ["n_chans", "n_outputs", "n_times"], None),
     ("SCCNet", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=250)),
     ("FBCNet", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=200)),
+    ("FBLightConvNet", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=200)),
     ("IFNetV2", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=200)),
     ("EEGMiner", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=200)),
     ("FBMSNet", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=200)),