+++ test/unit_tests/models/test_integration.py
@@ -17,7 +17,7 @@ from skorch.dataset import ValidSplit
 from braindecode.models.util import models_dict, models_mandatory_parameters
 from braindecode import EEGClassifier
 
-from braindecode.models import SyncNet, EEGSimpleConv, EEGResNet, USleep, EEGInceptionMI, FBCNet, FBMSNet, EEGMiner
+from braindecode.models import SyncNet, EEGSimpleConv, EEGResNet, USleep, EEGInceptionMI, FBCNet, FBMSNet, EEGMiner, FBLightConvNet
 
 
 # Generating the channel info
@@ -331,7 +331,7 @@ def test_model_has_drop_prob_parameter(model_class):
     named 'drop_prob' or any parameter that starts with 'activation'.
     """
 
-    if model_class in [SyncNet, EEGSimpleConv, EEGResNet, USleep, FBCNet, EEGMiner, EEGInceptionMI, FBMSNet]:
+    if model_class in [SyncNet, EEGSimpleConv, EEGResNet, USleep, FBCNet, EEGMiner, EEGInceptionMI, FBMSNet, FBLightConvNet]:
         pytest.skip(
             f"Skipping {model_class} as not dropout layer")